Переносит createTemplates в компоненты
