--- src/view/event-destination-description.js
@@ -1,3 +1,0 @@
-export const createEventDestinationDescriptionTemplate = (description) => {
-  return `<p class="event__destination-description">${description}</p>`;
-};
