--- src/view/event-offer.js
@@ -1,9 +1,0 @@
-export const createEventOfferTemplate = ({title, price}) => {
-  return `
-    <li class="event__offer">
-      <span class="event__offer-title">${title}</span>
-      &plus;&euro;&nbsp;
-      <span class="event__offer-price">${price}</span>
-    </li>
-  `;
-};
