--- src/view/event-photo.js
@@ -1,3 +1,0 @@
-export const createEventPhotoTemplate = (photoUrl) => {
-  return `<img class="event__photo" src="${photoUrl}" alt="Event photo">`;
-};
