--- src/view/trip-cost.js
@@ -1,7 +1,0 @@
-export const createTripCostTemplate = (cost) => {
-  return `
-    <p class="trip-info__cost">
-      Total: &euro;&nbsp;<span class="trip-info__cost-value">${cost}</span>
-    </p>
-  `;
-};
