+++ src/view/trip-event.js
@@ -2,7 +2,16 @@ import Component from './component';
 
 import dayjs from 'dayjs';
 import {toFormatTimeDiff} from './../utils/datetime';
-import {createEventOfferTemplate} from './event-offer';
+
+const createEventOfferTemplate = ({title, price}) => {
+  return (
+    `<li class="event__offer">
+      <span class="event__offer-title">${title}</span>
+      &plus;&euro;&nbsp;
+      <span class="event__offer-price">${price}</span>
+    </li>`
+  );
+};
 
 const createTripEventTemplate = ({type, city, offers, price, datetime, isFavorite}) => {
   return (
