+++ src/view/trip-info-container.js
@@ -1,7 +1,28 @@
 import Component from './component';
+import dayjs from 'dayjs';
 
-import {createTripInfoTemplate} from './trip-info';
-import {createTripCostTemplate} from './trip-cost';
+const createTripInfoTemplate = (cities = ``, firstDay, lastDay) => {
+  const firstDate = dayjs(firstDay);
+  const lastDate = dayjs(lastDay);
+
+  const lastDateString = firstDate.get(`month`) !== lastDate.get(`month`) ? lastDate.format(`MMM DD`) : lastDate.format(`DD`);
+
+  return (
+    `<div class="trip-info__main">
+      <h1 class="trip-info__title">${cities}</h1>
+
+      <p class="trip-info__dates">${dayjs(firstDay).format(`MMM DD`)}&nbsp;&mdash;&nbsp;${lastDateString}</p>
+    </div>`
+  );
+};
+
+const createTripCostTemplate = (cost) => {
+  return (
+    `<p class="trip-info__cost">
+      Total: &euro;&nbsp;<span class="trip-info__cost-value">${cost}</span>
+    </p>`
+  );
+};
 
 const createTripInfoContainerTemplate = (events) => {
   const cost = events.reduce((fullSum, event) => {
